openai/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
flatbuffers==25.9.23
griffe==1.15.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jh2==5.0.10
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.26.0
niquests==3.17.0
numpy==2.4.2
openai==2.16.0
openai-agents==0.7.0
openmeteo_requests==1.7.5
openmeteo_sdk==1.25.0
pandas==3.0.0
platformdirs==4.5.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
qh3==1.5.6
referencing==0.37.0
requests==2.32.5
requests-cache==1.3.0
retry-requests==2.0.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.52.1
tqdm==4.67.2
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.15.902
uvicorn==0.40.0
wassima==2.0.4
